move view_values to store
